--- a/javascript/net/grpc/web/grpcwebclientbase_test.js
+++ b/javascript/net/grpc/web/grpcwebclientbase_test.js
@@ -240,6 +240,13 @@ class MockXhr {
   }
 
   /**
+   * @param {number} ms
+   */
+  setTimeoutInterval(ms) {
+    return;
+  }
+
+  /**
    * @param {boolean} withCredentials
    */
   setWithCredentials(withCredentials) {