--- a/net/grpc/gateway/docker/prereqs/Dockerfile
+++ b/net/grpc/gateway/docker/prereqs/Dockerfile
@@ -12,39 +12,54 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-FROM grpcweb/common
 
-ARG MAKEFLAGS=-j8
+# Using multi-stage buid (See: https://docs.docker.com/develop/develop-images/multistage-build/)
+
+######################################
+# Stage 1: Fetch binaries
+######################################
+# node:... Docker image is based on buildpack-deps:stretch
+FROM buildpack-deps:stretch AS prepare
+
 ARG BUILDIFIER_VERSION=1.0.0
-ARG BAZEL_VERSION=4.1.0
+ARG PROTOBUF_VERSION=3.17.3
 
-RUN echo "\nloglevel=error\n" >> $HOME/.npmrc
+RUN apt-get -qq update && apt-get -qq install -y curl unzip
 
-WORKDIR /github/grpc-web
+WORKDIR /tmp
 
-# Check out scripts and initialize git submodules first to take advantage of docker build cache.
-COPY ./scripts ./scripts
-RUN ./scripts/init_submodules.sh
-
-COPY ./packages ./packages
-RUN cd ./packages/grpc-web && \
-  npm install && \
-  npm run build && \
-  npm link
-
-COPY ./Makefile ./Makefile
-COPY ./WORKSPACE ./WORKSPACE
-COPY ./bazel ./bazel
-COPY ./javascript ./javascript
-COPY ./net ./net
-COPY ./src ./src
-COPY ./test ./test
+RUN curl -sSL https://github.com/protocolbuffers/protobuf/releases/download/v$PROTOBUF_VERSION/\
+protoc-$PROTOBUF_VERSION-linux-x86_64.zip -o protoc.zip && \
+  unzip -qq protoc.zip && \
+  cp ./bin/protoc /usr/local/bin/protoc
 
 RUN wget -nv -O buildifier \
   https://github.com/bazelbuild/buildtools/releases/download/$BUILDIFIER_VERSION/buildifier && \
   chmod +x ./buildifier && \
-  ./buildifier --mode=check --lint=warn --warnings=all -r ./WORKSPACE bazel javascript net && \
-  rm ./buildifier
+  cp ./buildifier /usr/local/bin/buildifier
+
+# Download third_party modules to be used for the next stage
+WORKDIR /github/grpc-web
+
+RUN git clone https://github.com/grpc/grpc-web .
+COPY ./scripts/init_submodules.sh ./scripts/
+RUN ./scripts/init_submodules.sh
+
+
+######################################
+# Stage 2: Copy source files and build
+######################################
+FROM node:12.22.6-stretch AS copy-and-build
+
+ARG MAKEFLAGS=-j8
+ARG BAZEL_VERSION=4.1.0
+
+RUN mkdir -p /var/www/html/dist
+RUN echo "\nloglevel=error\n" >> $HOME/.npmrc
+
+COPY --from=prepare /usr/local/bin/protoc /usr/local/bin/
+COPY --from=prepare /usr/local/bin/buildifier /usr/local/bin/
+COPY --from=prepare /github/grpc-web/third_party /github/grpc-web/third_party
 
 RUN wget -nv -O bazel-installer.sh \
   https://github.com/bazelbuild/bazel/releases/download/$BAZEL_VERSION/\
@@ -53,6 +68,32 @@ bazel-$BAZEL_VERSION-installer-linux-x86_64.sh && \
   ./bazel-installer.sh && \
   rm ./bazel-installer.sh
 
-RUN bazel build javascript/net/grpc/web/... && \
+WORKDIR /github/grpc-web
+
+# Copy only files necessary to build the protoc-gen-grpc-web first as an optimization because they
+# are rarely updated compared with the javascript files.
+COPY ./WORKSPACE ./WORKSPACE
+COPY ./bazel ./bazel
+COPY ./javascript/net/grpc/web/grpc_generator.cc javascript/net/grpc/web/grpc_generator.cc
+COPY ./javascript/net/grpc/web/BUILD.bazel javascript/net/grpc/web/BUILD.bazel
+
+RUN bazel build javascript/net/grpc/web:protoc-gen-grpc-web && \
   cp $(bazel info bazel-genfiles)/javascript/net/grpc/web/protoc-gen-grpc-web \
   /usr/local/bin/protoc-gen-grpc-web
+
+COPY ./javascript ./javascript
+COPY ./packages ./packages
+
+RUN cd ./packages/grpc-web && \
+  npm install && \
+  npm run build && \
+  npm link
+
+COPY ./Makefile ./Makefile
+COPY ./net ./net
+COPY ./scripts ./scripts
+COPY ./src ./src
+COPY ./test ./test
+
+RUN /usr/local/bin/buildifier \
+  --mode=check --lint=warn --warnings=all -r ./WORKSPACE bazel javascript net