--- a/scripts/docker-run-jsunit-tests.sh
+++ b/scripts/docker-run-jsunit-tests.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Copyright 2018 Google LLC
+# Copyright 2021 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,14 +15,8 @@
 set -ex
 
 # This script is intended to be run within the base image from
-# net/grpc/gateway/docker/prereqs/Dockerfile
+# packages/grpc-web/docker/jsunit-test/Dockerfile
 
-cd /github/grpc-web && \
-  bazel clean && \
-  bazel test --cache_test_results=no \
-    //javascript/net/grpc/web/... \
-    //net/grpc/gateway/examples/...
+cd /grpc-web/packages/grpc-web
 
-cd /github/grpc-web/packages/grpc-web && \
-  npm run prepare && \
-  npm test
+npm run test-jsunit