--- a/scripts/test-proxy.sh
+++ b/scripts/test-proxy.sh
@@ -47,4 +47,6 @@ echo "$s1" | base64 -d | xxd
 # Take the 28 bytes we cut out above, the base64-encoded string should be this
 if [[ "$s1" != "AAAAAAcKBWhlbGxvgGdycGMtc3RhdHVzOjANCg==" ]]; then
   exit 1;
+else
+  echo "Envoy proxy test successful!"
 fi