### Key changes:

* Removed dependency on [Bazel](https://bazel.build/) / [rules_closure](https://github.com/bazelbuild/rules_closure) for Closure JsUnit tests in favor of [google-closure-deps](https://www.npmjs.com/package/google-closure-deps) and [Protractor](https://www.protractortest.org/#/).
  * This allow us the direct use use most up-to-date version of [Closure Library](https://github.com/google/closure-library) (v.s. [indirectly used](https://github.com/bazelbuild/rules_closure/blob/321e678c64656f9dfc39e978a0a017d2930eb7e8/closure/repositories.bzl#L709-L720) through rules_closure), and unblock code sync and [new releases](https://github.com/grpc/grpc-web/issues/1043).
    * Also ensures **tests are ran on the same version of Closure as release** (which was not before).
  * Protractor setup uses generated test HTML files, and is HEAVILY inspired by [FirebaseUi](https://github.com/firebase/firebaseui-web) (mainly just Dockerizing their [solution](https://github.com/firebase/firebaseui-web/tree/master/buildtools)) :)


* Optimized Test flows
  * Restructured Dockerfile layouts and test scripts so that unit tests runs earlier and **much faster!**
    * e.g. when user makes a javascript changes, reruns will **fail in as early as 10 seconds** (on jsunit failures) and at most in ~60 seconds (on mocha test failures), **compared to 4+ minutes as before** ! (on 2019 MacbookPro)
  * `npm test` in `packages/grpc-web` now runs both JsUnit tests on top of Mocha tests.

* Optimized Dockerfiles
  * Removed `common` Docker images (saves disk space) and revamped `prereqs` to use  [multistage](https://docs.docker.com/develop/develop-images/multistage-build/) build.
    * Avoids the  possibility code from previous `git clone` to be accidentally in local builds
  * Optimized `prereqs` for faster build
    * Reduced build targets so `prereqs` build time is down from 5+ minutes to <3 minutes (on 2019 MacbookPro)
    * Optimized build sequence so larger / less frequently changed targets are built earlier to best leverage the Docker [build cache](https://pythonspeed.com/articles/docker-caching-model/).

### Other (small) changes:

* Fix small flake in interop test where it sometimes times out at 500ms (updated timeout to 1 second)
  * .. and updated mocha unit tests timeout to 10 seconds
* Cleaned up echo example Makefile to ignore testing files (rather than removing then adding them back)
* Created `.dockerignore` to avoid build failures when local node_modules/ are present
* Deleted `streambodyclientreadablestream.js` since it’s not used and doesn’t build.

### Follow-up:

* Clean up remaining `closure_js_library` usages and our [bazel rules](https://github.com/grpc/grpc-web/tree/master/bazel).